course work/coursach/requirements.txt
Django>=4.2,<5.0
django-simple-history>=2.0,<3.0
django-import-export>=3.0,<4.0
drf-yasg>=1.20,<2.0
djangorestframework>=3.14,<4.0
django-filter>=24.3,<25.0
celery>=5.2,<6.0